# openapi_updated (1).yaml
openapi: 3.0.1
info:
  contact: {}
  description: Api Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  termsOfService: urn:tos
  title: Api Documentation
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /ads:
    get:
      tags:
        - Объявления
      operationId: getALLAds
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseWrapperAdDto'
    post:
      tags:
        - Объявления
      summary: addAds
      operationId: addAds
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - image
                - properties
              type: object
              properties:
                properties:
                  $ref: '#/components/schemas/CreateAdDto'
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdDto'
        '404':
          description: Not Found
        '403':
          content: {}
          description: Forbidden
        '401':
          content: {}
          description: Unauthorized
  /ads/{ad_pk}/comments:
    get:
      tags:
        - Объявления
      summary: getComments
      operationId: getComments
      parameters:
        - name: ad_pk
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseWrapperCommentDto'
        '404':
          description: Not Found

    post:
      tags:
        - Объявления
      summary: addComments
      operationId: addComments
      parameters:
        - name: ad_pk
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentDto'
        '404':
          description: Not Found
        '403':
          content: {}
          description: Forbidden
        '401':
          content: {}
          description: Unauthorized
...
  /ads/{id}:
    get:
      tags:
        - Объявления
      summary: getFullAd
      operationId: getAds
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FullAdDto'
        '404':
          description: Not Found
    delete:
      tags:
        - Объявления
      summary: removeAds
      operationId: removeAds
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          content: {}
          description: No Content
        '401':
          content: {}
          description: Unauthorized
        '403':
          content: {}
          description: Forbidden
    patch:
      tags:
        - Объявления
      summary: updateAds
      operationId: updateAds
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdDto'
        '404':
          description: Not Found
        '403':
          content: {}
          description: Forbidden
        '401':
          content: {}
          description: Unauthorized
  /ads/{ad_pk}/comments/{id}:
    get:
      tags:
        - Объявления
      summary: getComments
      operationId: getComments_1
      parameters:
        - name: ad_pk
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentDto'
        '404':
          description: Not Found
    delete:
      tags:
        - Объявления
      summary: deleteComments
      operationId: deleteComments
      parameters:
        - name: ad_pk
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
        '403':
          content: {}
          description: Forbidden
        '401':
          content: {}
          description: Unauthorized
    patch:
      tags:
        - Объявления
      summary: updateComments
      operationId: updateComments
      parameters:
        - name: ad_pk
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentDto'
        '404':
          description: Not Found
        '403':
          content: {}
          description: Forbidden
        '401':
          content: {}
          description: Unauthorized
  /ads/me:
    get:
      operationId: getAdsMeUsingGET
      parameters:
        - in: query
          name: authenticated
          schema:
            type: boolean
        - in: query
          name: authorities[0].authority
          schema:
            type: string
        - in: query
          name: credentials
          schema:
            type: object
        - in: query
          name: details
          schema:
            type: object
        - in: query
          name: principal
          schema:
            type: object
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseWrapperAdDto'
          description: OK
        '401':
          content: {}
          description: Unauthorized
        '403':
          content: {}
          description: Forbidden
        '404':
          content: {}
          description: Not Found
      summary: getAdsMe
      tags:
        - Объявления
      x-accepts: '*/*'
  /image/{id}:
    patch:
      tags:
        - Изображения
      summary: updateAdsImage
      operationId: updateImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: array
                items:
                  type: string
                  format: byte
        '404':
          description: Not Found
...